+++ ModelSwapperSkins/ModelInfos/ModelInfoInitializer.cs
@@ -131,6 +131,7 @@ namespace ModelSwapperSkins.ModelInfos
             SetModelInfo("ImpBossBody", new ModelInfo(12.5f));
             SetModelInfo("JellyfishBody", new ModelInfo(7.5f));
             SetModelInfo("LemurianBody", new ModelInfo(20f));
+            SetModelInfo("LemurianBruiserBody", new ModelInfo(25f));
             SetModelInfo("LunarExploderBody", new ModelInfo(5f));
             SetModelInfo("RailgunnerBody", new ModelInfo(1.5f));
             SetModelInfo("ToolbotBody", new ModelInfo(20f));